Update error handling and macro to work with prismatic 0.2.0
In prismatic plumbing 0.2.0, there are a few noteworthy changes:

* They have ported over to use the `schema` library for managing
  fnk schemas, and
* They accepted a pull request to use ex-info to give more interrogable
  exceptions when a graph fails to compile or instantiate.

This branch ports us over to be compatible with both of those changes.
I don't want to have a dependency on a SNAPSHOT version of plumbing,
so we'll have to wait until they do a release before we can actually
merge this.