Scratch file showing how we might modify reification
This commit includes a file that has some sandbox code that shows
a possible alternative to how we currently handle our service
reification.  The idea is that instead of doing the reification
as a wrapper around the prismatic graph functions, we would do
the prismatic graph functions as a wrapper around the reified
service instances.  This would allow the `this` keyword to
be used a lot more flexibly inside of `defservice` statements,
and would potentially allow us to get rid of a lot of the
magic that the macro currently performs in manipulating
the contents of a `defservice` statement.