(TK-291) Handle all events in testutils.logging
Rework testutils.logging to tie in at the logback (Java) level, rather
than the clojure.tools.logging level, allowing us to see and manipulate
most, if not all log events.

Have the core API handle only native event objects for efficiency, but
provide a more friendly map representation via event->map.

Add a set of scoped call-with-* and with-* handlers to allow temporarily
overriding the logging appenders, level, etc.

Rework the existing with-test-logging, with-test-logging-debug, and
logged? code to use the new API, add some tests, and add some minor new,
backward-compatible features.  Using the new API does change the
fundamental semantics a bit, since logged? may see more events than it
used to (perhaps from ActiveMQ or...), but that seems unlikely to break
existing tests unless the the selectors weren't specific enough to begin
with, and is easily fixed.

Add a new with-log-suppressed-unless-notable macro that suppresses all
logging unless an event matches its notable-event? predicate, in which
case it will dump the log to *err* and provide a path to the temporary
log file.

Deprecate much of the previous code, which should for the most part be
superceded by the new calls.

cf. PDB-1995

[This began as PuppetDB 5a4789fcd3a82514a335ab59735965ef0d2a6e59]