(MAINT) provide schema for app-context
This commit adds a schema for TrapperkeeperAppContext, which
allows us to clean up a bunch of precondition assertions and
establish a better contract for what makes up an app context.
This will be important for some subsequent commits where I will
be adding more data to the context.

Note that as part of this commit, I also moved all of the service
contexts for the individual services out of the root of the
app-context map, where they previously existed at the same level
as the other well-known keys of the app-context map.  Now they
live in their own nested map with a key called `:service-contexts`,
which makes the schema for app-context a lot cleaner.